better organize testing and plotting utils
