+++ scripts/test_eval/tester.py
@@ -1,16 +1,18 @@
-from log_parser import LogParser, DEFAULT_LOG_FOLDER
+from .logparser import LogParser, DEFAULT_LOG_FOLDER
 from subprocess import PIPE, run
 import argparse
 import os
 import pandas as pd
 import uuid
 
-PASH_TOP = "/home/tamlu/pash"
-TESTFILES = [f"{PASH_TOP}/evaluation/scripts/input/1M.txt", f"{PASH_TOP}/evaluation/scripts/input/10M.txt", f"{PASH_TOP}/evaluation/scripts/input/100M.txt", f"{PASH_TOP}/evaluation/scripts/input/1G.txt", f"{PASH_TOP}/evaluation/scripts/input/3G.txt", f"{PASH_TOP}/evaluation/scripts/input/100G.txt"]
-BATCHSZ = [10000, 100000, 1000000, 10000000, 30000000, 100000000]
+GIT_TOP_CMD = [ 'git', 'rev-parse', '--show-toplevel', '--show-superproject-working-tree']
+if 'PASH_TOP' in os.environ:
+    PASH_TOP = os.environ['PASH_TOP']
+else:
+    PASH_TOP = run(GIT_TOP_CMD, stdout=PIPE, stderr=PIPE, universal_newlines=True).stdout.rstrip()
 
 class Tests(LogParser):
-    def __init__(self, in_file = TESTFILES[1], batch_sz = 100000):
+    def __init__(self, in_file = None, batch_sz = 100000):
         self.in_file = in_file
         self.batch_sz = str(batch_sz)
         self.log_parser = LogParser()
@@ -35,7 +37,8 @@ class Tests(LogParser):
             in_file = self.in_file
 
         new_env["IN"] = in_file
-
+        new_env["PASH_TOP"] = PASH_TOP
+        
         command = [f"{PASH_TOP}/pa.sh", test_path, "--output_time", f"-w {width}", "-d 1"]
 
         if r_split:
@@ -87,23 +90,3 @@ class Tests(LogParser):
         temp_filename = os.path.basename(test_path).replace(".sh", "") + "_" + str(uuid.uuid4()) + ".log"
         log_file = os.path.join(log_folder, temp_filename)
         return log_file
-
-
-def run_tests():
-    test10M = Tests(TESTFILES[3], BATCHSZ[3])
-    tests = ['minimal_sort', 'topn', 'wf', 'diff', 'set-diff', 'double_sort', 'sort'] #'bigrams', 'spell' 'shortest_scripts'
-    tests = ["double_sort", "diff"]
-    test_files = []
-    for test in tests:
-        test_path = f"{PASH_TOP}/evaluation/microbenchmarks/{test}.sh"
-        test_files.append(test_path)
-    tmp_folder = "tmp2_1G_width4rm/"
-    test10M.run_test_list(test_files, r_split=True, width=4, log_folder=tmp_folder)
-    test10M.run_test_list(test_files, r_split=True, width=4, no_eager=True, log_folder=tmp_folder)
-    test10M.run_test_list(test_files, width=4, log_folder=tmp_folder)
-    test10M.run_test_list(test_files, width=4, no_eager=True, log_folder=tmp_folder)
-    print(test10M.get_df()[["test_name", "no_eager", "split_type", "exec_time", "cpu%", "width"]].to_string(index = False))
-
-if __name__ == '__main__':
-    os.environ["PASH_TOP"] = PASH_TOP
-    run_tests()
